+++ libraries/decoder_opus/src/main/java/androidx/media3/decoder/opus/LibopusAudioRenderer.java
@@ -31,6 +31,7 @@ import androidx.media3.exoplayer.audio.AudioSink;
 import androidx.media3.exoplayer.audio.AudioSink.SinkFormatSupport;
 import androidx.media3.exoplayer.audio.DecoderAudioRenderer;
 import androidx.media3.extractor.VorbisUtil;
+import com.google.common.primitives.ImmutableIntArray;
 
 /** Decodes and renders audio using the native Opus decoder. */
 @UnstableApi
@@ -134,7 +135,7 @@ public class LibopusAudioRenderer extends DecoderAudioRenderer<OpusDecoder> {
 
   @Nullable
   @Override
-  protected int[] getChannelMapping(OpusDecoder decoder) {
+  protected ImmutableIntArray getChannelMapping(OpusDecoder decoder) {
     return VorbisUtil.getVorbisToAndroidChannelLayoutMapping(decoder.channelCount);
   }